move PojoBeanMapper out, move errors to
